Rewrite the diff block list and buffer registry
`block.rs`. The registry half is small; the work is `diff_mark_adjust_tp`,
190 lines that keep a tabpage's block list correct across an edit without
recomputing it. Upstream numbers its six cases in a diagram comment and
then writes them as one nest of conditions; the numbers are now on the
arms, and each says what the case *is*.

Two shapes the transpile hid. The `check_unchanged` flag exists because
the trim has to run **after** the other buffers' ranges are adjusted --
it compares the block's lines across all of them -- so it is a flag and
not a call at the site that sets it. And the three-way
`inserted`/`deleted` decision at the top is a free function with the three
`mark_adjust` call shapes it decodes named in its arms.

`adjust_others` and `merge_or_advance` are closures, not functions: one
call site each, and a closure written inside an `unsafe` block inherits
it, so the split costs the ratchet nothing (`traps-ratchet.md`).
`merge_or_advance` answers the new `(dprev, dp)` pair rather than taking
two `&mut *mut`, which is what makes it a closure at all.

`diff_check_unchanged`'s two passes are `for dir in [FORWARD, BACKWARD]`,
which is what the C's `for(;;) { ...; if (dir == BACKWARD) break; dir =
BACKWARD; }` was; `off_org`/`off_new` were declared outside that loop and
only ever assigned in the backward pass, so they are now locals with the
direction in their initialiser.

`diff_buf_adjust`'s "is any other window still showing this buffer in
diff mode" search early-returns instead of setting a flag, and
`get_max_diff_length`/`valid_diff`/`diff_mode_buf`/`diff_buf_idx` are
each an iterator expression. The doc says what each is *for*, which for
`valid_diff` is the reason it exists: `:diffget` runs autocommands
between reading a block and using it.

Gates: diffsweep byte-identical; `oldtest test_diffmode` 83 executed, 0
failed; clippy 2,897 -> 2,896; ratchet unmoved at 18.

+++ crates/nvim/src/nvim/diff/hunk.rs
@@ -146,7 +146,7 @@ unsafe extern "C" fn process_hunk(
                 > (*dp).df_lnum[idx_orig as usize] + (*dp).df_count[idx_orig as usize]
         {
             if *notsetp {
-                diff_copy_entry(dprev, dp, idx_orig, idx_new);
+                diff_copy_entry(dprev, dp, idx_orig as usize, idx_new as usize);
             }
             dprev = dp;
             dp = (*dp).df_next;
@@ -238,7 +238,7 @@ unsafe extern "C" fn process_hunk(
             let mut i_1: ::core::ffi::c_int = idx_orig + 1 as ::core::ffi::c_int;
             while i_1 < idx_new {
                 if !(*curtab.get()).tp_diffbuf[i_1 as usize].is_null() {
-                    diff_copy_entry(dprev, dp, idx_orig, i_1);
+                    diff_copy_entry(dprev, dp, idx_orig as usize, i_1 as usize);
                 }
                 i_1 += 1;
             }
@@ -301,7 +301,7 @@ pub(crate) unsafe extern "C" fn diff_read(
         }
         while !dp.is_null() {
             if notset {
-                diff_copy_entry(dprev, dp, idx_orig, idx_new);
+                diff_copy_entry(dprev, dp, idx_orig as usize, idx_new as usize);
             }
             dprev = dp;
             dp = (*dp).df_next;
